requirements.txt
tqdm==4.67.1
pyglet==2.1.6
ipdb
h5py==3.11.0
pyyaml==6.0.2
tensorboard==2.14.0
matplotlib==3.7.5
opencv-python==4.12.0.88
pycollada==0.9.2
scipy==1.10.1
trimesh==4.7.1
wandb
rl-games
warp-lang